remove paper module
