+++ .github/workflows/build-commit.yml
@@ -5,7 +5,7 @@ name: build-commit
 on:
   push:
     paths-ignore:
-      - "readme.md"
+      - "README.md"
       - "LICENSE"
       - ".gitignore"
       - "changelog.md"
@@ -13,7 +13,7 @@ on:
       - "src/*/resources/lang/*"
   pull_request:
     paths-ignore:
-      - "readme.md"
+      - "README.md"
       - "LICENSE"
       - ".gitignore"
       - "changelog.md"
@@ -38,7 +38,17 @@ jobs:
       - name: Setup Gradle
         uses: gradle/actions/setup-gradle@v3
       - name: Build with Gradle
-        run: ./gradlew :telekinesis-fabric:buildAllVersions --stacktrace
+        run: ./gradlew buildAllVersions --stacktrace
       - uses: actions/upload-artifact@v4
         with:
           path: versions/**/build/libs/*.jar
+
+  modrinth-description:
+    runs-on: ubuntu-latest
+    name: Sync Modrinth description
+    steps:
+      - uses: actions/checkout@v4
+      - uses: funnyboy-roks/modrinth-auto-desc@v1.6
+        with:
+          auth-token: ${{ secrets.MODRINTH_API_KEY }}
+          slug: 'skylper'
