+++ .github/workflows/release.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Setup Gradle
         uses: gradle/actions/setup-gradle@v3
       - name: Release with Gradle
-        run: ./gradlew :telekinesis-fabric:releaseAllVersions --stacktrace
+        run: ./gradlew releaseAllVersions postUpdate --stacktrace
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           MODRINTH_API_KEY: ${{ secrets.MODRINTH_API_KEY }}
@@ -31,3 +31,13 @@ jobs:
           ORG_GRADLE_PROJECT_signingKey: ${{ secrets.GPG_SECRET_KEY }}
           ORG_GRADLE_PROJECT_signingPassword: ${{ secrets.GPG_SECRET_KEY_PASSWORD }}
           GPG_SECRET_KEY_ID: ${{ secrets.GPG_SECRET_KEY_ID }}
+
+  modrinth-description:
+    runs-on: ubuntu-latest
+    name: Sync Modrinth description
+    steps:
+      - uses: actions/checkout@v4
+      - uses: funnyboy-roks/modrinth-auto-desc@v1.6
+        with:
+          auth-token: ${{ secrets.MODRINTH_API_KEY }}
+          slug: 'skylper'
