+++ src/main/java/dev/nyon/telekinesis/mixins/StriderMixin.java
