--- telekinesis-fabric/build.gradle.kts
@@ -1,203 +1,0 @@
-@file:Suppress("SpellCheckingInspection", "SENSELESS_COMPARISON")
-
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    kotlin("jvm") version "2.0.0"
-    kotlin("plugin.serialization") version "2.0.0"
-    id("fabric-loom") version "1.7-SNAPSHOT"
-
-    id("me.modmuss50.mod-publish-plugin") version "0.5.+"
-
-    `maven-publish`
-    signing
-}
-
-val beta: Int = 1 // Pattern is '1.0.0-beta1-1.20.6-pre.2'
-val featureVersion = "1.0.0${if (beta != null) "-beta$beta" else ""}"
-val mcVersion = property("mcVersion")!!.toString()
-val mcVersionRange = property("mcVersionRange")!!.toString()
-version = "$featureVersion-$mcVersion"
-
-group = "dev.nyon"
-val projectAuthors = listOf("btwonion")
-val githubRepo = "btwonion/telekinesis"
-
-base {
-    archivesName.set(rootProject.name)
-}
-
-loom {
-    if (stonecutter.current.isActive) {
-        runConfigs.all {
-            ideConfigGenerated(true)
-            runDir("../../run")
-        }
-    }
-
-    mixin { useLegacyMixinAp = false }
-}
-
-// Enable data generation for >1.20.6
-if (!listOf("1.20.1", "1.20.4", "1.20.6").contains(stonecutter.current.version)) {
-    fabricApi {
-        configureDataGeneration()
-    }
-}
-
-repositories {
-    mavenCentral()
-    exclusiveContent {
-        forRepository {
-            maven("https://api.modrinth.com/maven")
-        }
-        filter {
-            includeGroup("maven.modrinth")
-        }
-    }
-    maven("https://maven.terraformersmc.com")
-    maven("https://maven.parchmentmc.org")
-    maven("https://repo.nyon.dev/releases")
-    maven("https://maven.isxander.dev/releases")
-    maven("https://maven.isxander.dev/snapshots")
-    maven("https://jitpack.io")
-}
-
-dependencies {
-    minecraft("com.mojang:minecraft:$mcVersion")
-    mappings(loom.layered {
-        val parchment: String = property("deps.parchment").toString()
-        if (parchment.isNotEmpty()) parchment("org.parchmentmc.data:parchment-$parchment@zip")
-        officialMojangMappings()
-    })
-
-    implementation("org.vineflower:vineflower:1.10.1")
-    modImplementation("net.fabricmc:fabric-loader:0.15.11")
-    modImplementation("net.fabricmc.fabric-api:fabric-api:${property("deps.fapi")!!}")
-    modImplementation("net.fabricmc:fabric-language-kotlin:1.11.0+kotlin.2.0.0")
-
-    modImplementation("dev.isxander:yet-another-config-lib:${property("deps.yacl")!!}")
-    modImplementation("com.terraformersmc:modmenu:${property("deps.modMenu")!!}")
-
-    include(implementation(annotationProcessor("com.github.bawnorton.mixinsquared:mixinsquared-fabric:0.1.1")!!)!!)
-
-    include(modImplementation("dev.nyon:konfig:2.0.1-1.20.4")!!)
-}
-
-val javaVersion = property("javaVer")!!.toString()
-tasks {
-    processResources {
-        val modId = "telekinesis"
-        val modName = "telekinesis"
-        val modDescription = "Adds a telekinesis enchantment to minecraft"
-
-        val props = mapOf(
-            "id" to modId,
-            "name" to modName,
-            "description" to modDescription,
-            "version" to project.version,
-            "github" to githubRepo,
-            "mc" to mcVersionRange
-        )
-
-        props.forEach(inputs::property)
-
-        filesMatching("fabric.mod.json") {
-            expand(props)
-        }
-    }
-
-    register("releaseMod") {
-        group = "publishing"
-
-        dependsOn("publishMods")
-        dependsOn("publish")
-    }
-
-    withType<JavaCompile> {
-        options.release = javaVersion.toInt()
-    }
-
-    withType<KotlinCompile> {
-        compilerOptions {
-            jvmTarget = JvmTarget.fromTarget(javaVersion)
-        }
-    }
-}
-
-val changelogText = buildString {
-    append("# v${project.version}\n")
-    rootProject.file("changelog.md").readText().also(::append)
-}
-
-val supportedMcVersions: List<String> =
-    property("supportedMcVersions")!!.toString().split(',').map(String::trim).filter(String::isNotEmpty)
-
-publishMods {
-    displayName = "v${project.version}"
-    file = tasks.remapJar.get().archiveFile
-    changelog = changelogText
-    type = if (beta != null) BETA else STABLE
-    modLoaders.addAll("fabric", "quilt")
-
-    modrinth {
-        projectId = "LLfA8jAD"
-        accessToken = providers.environmentVariable("MODRINTH_API_KEY")
-        minecraftVersions.addAll(supportedMcVersions)
-
-        requires { slug = "fabric-api" }
-        requires { slug = "yacl" }
-        requires { slug = "fabric-language-kotlin" }
-        optional { slug = "modmenu" }
-    }
-
-    github {
-        repository = githubRepo
-        accessToken = providers.environmentVariable("GITHUB_TOKEN")
-        commitish = "master"
-    }
-}
-
-publishing {
-    repositories {
-        maven {
-            name = "nyon"
-            url = uri("https://repo.nyon.dev/releases")
-            credentials {
-                username = providers.environmentVariable("NYON_USERNAME").orNull
-                password = providers.environmentVariable("NYON_PASSWORD").orNull
-            }
-        }
-    }
-    publications {
-        create<MavenPublication>("maven") {
-            groupId = "dev.nyon"
-            artifactId = "telekinesis"
-            version = project.version.toString()
-            from(components["java"])
-        }
-    }
-}
-
-java {
-    withSourcesJar()
-
-    javaVersion.toInt().let { JavaVersion.values()[it - 1] }.let {
-        sourceCompatibility = it
-        targetCompatibility = it
-    }
-}
-
-/*
-signing {
-    val signingKey: String? by project
-    val signingPassword: String? by project
-    useGpgCmd()
-    if (signingKey != null && signingPassword != null) {
-        useInMemoryPgpKeys(signingKey, signingPassword)
-    }
-    sign(publishing.publications)
-}
- */
-
