--- telekinesis-paper/src/main/java/dev/nyon/telekinesis/PaperLoader.java
@@ -1,22 +1,0 @@
-package dev.nyon.telekinesis;
-
-import io.papermc.paper.plugin.loader.PluginClasspathBuilder;
-import io.papermc.paper.plugin.loader.PluginLoader;
-import io.papermc.paper.plugin.loader.library.impl.MavenLibraryResolver;
-import org.eclipse.aether.artifact.DefaultArtifact;
-import org.eclipse.aether.graph.Dependency;
-import org.eclipse.aether.repository.RemoteRepository;
-import org.jetbrains.annotations.NotNull;
-
-@SuppressWarnings("UnstableApiUsage")
-public class PaperLoader implements PluginLoader {
-    @Override
-    public void classloader(@NotNull PluginClasspathBuilder classpathBuilder) {
-        MavenLibraryResolver resolver = new MavenLibraryResolver();
-        resolver.addDependency(new Dependency(new DefaultArtifact("org.jetbrains.kotlin:kotlin-stdlib:1.9.0"), null));
-        resolver.addDependency(new Dependency(new DefaultArtifact("com.akuleshov7:ktoml-core-jvm:0.5.0"), null));
-        resolver.addRepository(new RemoteRepository.Builder("central", "default", "https://repo1.maven.org/maven2/").build());
-
-        classpathBuilder.addLibrary(resolver);
-    }
-}
