+++ versions/1.20.6/gradle.properties
@@ -2,7 +2,7 @@ mcVersion=1.20.6
 mcVersionRange=>=1.20.5 <=1.20.6
 supportedMcVersions=1.20.5,1.20.6
 
-deps.parchment=1.20.6:2024.05.01
+deps.parchment=
 deps.fapi=0.100.0+1.20.6
 deps.yacl=3.4.2+1.20.5-fabric
 deps.modMenu=10.0.0-beta.1
